Refactor edvorakjp user library
