+++ users/edvorakjp/readme.md
@@ -10,7 +10,7 @@ This is a sample. You can swap any symbol keys and modifier keys.
   //+----+----+----+----+----+----+----+----+----+----+----+----+----+---------+
       `  , !  , @  , #  , $  , %  , ^  , &  , *  , (  , )  , [  , ]  ,  BSPC   ,
   //+----+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+-------+
-      TAB  , '  , ,  , .  , P  , Y  , F  , G  , R  , W  , Q  , /  , =  ,   \   ,
+      TAB  , '  , ,  , .  , Y  , P  , F  , G  , R  , W  , Q  , /  , =  ,   \   ,
   //+------++---++---++---++---++---++---++---++---++---++---++---++---+-------+
       CAPS  , A  , O  , E  , I  , U  , D  , T  , N  , S  , M  , -  ,    ENT    ,
   //+-------+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-----------+
@@ -25,7 +25,7 @@ This is a sample. You can swap any symbol keys and modifier keys.
   //+----+----+----+----+----+----+----+----+----+----+----+----+----+---------+
       `  , !  , @  , #  , $  , %  , ^  , &  , *  , (  , )  , [  , ]  ,  BSPC   ,
   //+----+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+-------+
-      TAB  , '  , ,  , .  , P  , Y  , F  , G  , R  , W  , C  , /  , =  ,
+      TAB  , '  , ,  , .  , Y  , P  , F  , G  , R  , W  , C  , /  , =  ,
   //+------++---++---++---++---++---++---++---++---++---++---++---++---++
       CAPS  , A  , O  , E  , I  , U  , D  , T  , N  , S  , M  , ;  , -  , ENT  ,
   //+-------+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+-+--+------+
