+++ users/edvorakjp/rules.mk
@@ -1,1 +1,3 @@
-SRC += edvorakjp.c
+SRC += edvorakjp.c \
+       edvorakjp_process_record.c \
+       edvorakjp_status.c
